Revert "Import Samsung release N910FXXU1BOB4 kernel source."
This reverts commit e00882fa2a467405d835d79728c4ed64d00b44ff.

Conflicts:
	arch/arm/configs/apq8084_sec_defconfig
	arch/arm/configs/apq8084_sec_trlte_eur_defconfig
	drivers/input/keyboard/cypress_touchkey_t/cypress-touchkey.c
	drivers/input/keyboard/cypress_touchkey_t/cypress-touchkey.h
	drivers/input/keyboard/gpio_keys.c
	drivers/media/isdbt/Makefile
	drivers/motor/tspdrv.c
	drivers/video/msm/mdss/Kconfig
	drivers/video/msm/mdss/Makefile
	drivers/video/msm/mdss/mdss_mdp.c
	scripts/fips_crypto_hmac.sh

+++ include/linux/time.h
@@ -68,7 +68,7 @@ static inline struct timespec timespec_add(struct timespec lhs,
 {
 	struct timespec ts_delta;
 	set_normalized_timespec(&ts_delta, lhs.tv_sec + rhs.tv_sec,
-				(s64)lhs.tv_nsec + rhs.tv_nsec);
+				lhs.tv_nsec + rhs.tv_nsec);
 	return ts_delta;
 }
 
@@ -80,7 +80,7 @@ static inline struct timespec timespec_sub(struct timespec lhs,
 {
 	struct timespec ts_delta;
 	set_normalized_timespec(&ts_delta, lhs.tv_sec - rhs.tv_sec,
-				(s64)lhs.tv_nsec - rhs.tv_nsec);
+				lhs.tv_nsec - rhs.tv_nsec);
 	return ts_delta;
 }
